pyproject.toml
[tool.poetry]
name = "helsing-solution"
version = "0.1.0"
description = ""
authors = ["calebgcc <calebgucciardi@gmail.com>"]
readme = "README.md"
packages = [{include = "helsing_solution"}]

[tool.poetry.dependencies]
python = "^3.10"
protobuf = "^6.30.2"
grpcio = "^1.71.0"
grpcio-tools = "^1.71.0"
googleapis-common-protos = "^1.70.0"
dotenv = "^0.9.9"


[tool.poetry.group.dev.dependencies]
mypy-protobuf = "^3.6.0"
grpcio-tools = "^1.71.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
